requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
openai==1.3.0
selenium==4.15.2
beautifulsoup4==4.12.2
requests==2.31.0
pdfplumber==0.10.3
python-docx==1.1.0
sqlalchemy==2.0.23
aiofiles==23.2.1
httpx==0.25.2

